Moving zedUpload implementation to eve
