+++ pkg/pillar/Gopkg.toml
@@ -59,10 +59,6 @@
 
 [[constraint]]
   branch = "master"
-  name = "github.com/zededa/shared"
-
-[[constraint]]
-  branch = "master"
   name = "github.com/zededa/api"
 
 [prune]
