+++ pkg/pillar/zedUpload/awsutil/driver.go
