+++ pkg/pillar/zedUpload/awsutil/s3api.go
