+++ pkg/pillar/zedUpload/awsutil/s3sync.go
