+++ pkg/pillar/zedUpload/azureutil/azure.go
