+++ pkg/pillar/zedUpload/datastore.go
