+++ pkg/pillar/zedUpload/datastore_req.go
