+++ pkg/pillar/zedUpload/httputil/http.go
